Moves Institution to a separate file
