+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationRequest/HearingArrangements/HearingArrangementsForm.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react'
 import { useIntl } from 'react-intl'
-import { Case, Institution } from '@island.is/judicial-system/types'
+import { Case } from '@island.is/judicial-system/types'
+import type { Institution } from '@island.is/judicial-system/types'
 import {
   FormContentContainer,
   FormFooter,
