+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/SelectCourt/SelectCourt.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { Box, Select, Text } from '@island.is/island-ui/core'
 import { ReactSelectOption } from '@island.is/judicial-system-web/src/types'
 import { setAndSendToServer } from '@island.is/judicial-system-web/src/utils/formHelper'
-import { Case, Institution } from '@island.is/judicial-system/types'
+import { Case } from '@island.is/judicial-system/types'
+import type { Institution } from '@island.is/judicial-system/types'
 import { ValueType } from 'react-select'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
