+++ libs/judicial-system/types/src/index.ts
@@ -1,7 +1,10 @@
 export { Feature } from './lib/feature'
+
+export { InstitutionType } from './lib/institution'
+export type { Institution } from './lib/institution'
+
 export * from './lib/types'
 export type {
-  Institution,
   User,
   CreateUser,
   UpdateUser,
